backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy>=2.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-multipart>=0.0.9
pyyaml>=6.0
tenacity>=9.0.0
hl7apy>=1.3.4
pymupdf>=1.24.0
